pyproject.toml
[project]
name = "veasy-peasy"
version = "0.1.0"
description = "Local-first CLI that scans visa application documents and extracts structured data"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "typer>=0.12",
    "pyyaml>=6",
    "passporteye>=2.2",
    "easyocr>=1.7",
    "pymupdf>=1.24",
]

[project.scripts]
veasy-peasy = "veasy_peasy.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/veasy_peasy"]

[dependency-groups]
dev = ["pytest"]
